refactor: admin components and actions to improve structure and organization (logs, messages, dashboard)
