+++ src/app/admin/messages/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PageTitle from '@/components/private/page-title';
 import PageAnimation from '@/components/public/page-animation';
-import MessagesContainer from '@/app/admin/messages/(components)/msg-container';
+import { MessageContainer } from '@/features/admin/messages/components';
 
 export const metadata = {  
   title: 'Customer Messages | Admin Dashboard',  
@@ -14,7 +14,7 @@ const Page = () => {
       <div className="p-6">
         <PageTitle title='Customer Messages' />
         <div className="space-y-4">
-          <MessagesContainer />
+          <MessageContainer />
         </div>
       </div>
     </PageAnimation>
