+++ src/features/admin/dashboard/actions/index.ts
@@ -1,0 +1,2 @@
+export { default as getDashboardStats } from './dashboard.actions';
+export { default as getAnnouncements } from './getAnnouncements';
