+++ src/features/admin/logs/components/index.ts
@@ -1,0 +1,3 @@
+export { LogDetailsModal } from './log-details-modal';
+export { LogsContent } from './logs-content';
+export { LogsTable } from './logs-table';
