+++ src/features/admin/messages/components/msg-detail-original.tsx
