+++ src/features/admin/messages/components/msg-detail-replyform.tsx
@@ -7,8 +7,8 @@ import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
 import { Form } from "@/components/ui/form";
-import { replyMessageSchema } from "@/schema/messages";
-import { replyToMessage } from "@/app/admin/messages/_actions";
+import { replyMessageSchema } from "@/features/admin/messages/messages.schema";
+import { replyToMessage } from "@/features/admin/messages/actions";
 import { useUserStore } from "@/stores/user.store";
 import useToast from "@/hooks/use-toast";
 import { cn } from "@/lib/utils";
